requirements.txt
# Vietnamese Story Audio to Video Generator
# Core dependencies

# Google Gemini API
google-generativeai>=0.3.0

# Image processing
Pillow>=10.0.0

# Video processing (optional, for fallback)
imageio>=2.31.0
imageio-ffmpeg>=0.4.8

# Audio processing
pydub>=0.25.1

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Development dependencies (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

